requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
coreapi==2.3.3
coreschema==0.0.4
cx-Oracle==8.3.0
Django==4.1
django-admin-interface==0.19.2
django-colorfield==0.7.2
django-cors-headers==3.13.0
django-filter==22.1
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-naomi==0.8
djangorestframework==3.13.1
drf-extra-fields==3.4.0
drf-yasg==1.21.3
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mysqlclient==2.1.1
packaging==21.3
Pillow==9.2.0
pyparsing==3.0.9
python-environ==0.4.54
pytz==2022.1
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.11
